requirements.txt
flask
requests
mysqlclient
